rename latency_predictor
